requirements.txt
uvicorn==0.23.2
fastapi==0.104.1
mysql-connector-python==8.2.0
python-multipart==0.0.6
aiofiles==23.2.1
passlib==1.7.4
python-jose==3.3.0